% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain-blfs.R
\name{check_files_blfs}
\alias{check_files_blfs}
\title{Identify large files that should be tracked with Box LFS}
\usage{
check_files_blfs(dir = NULL, size = 10, new = FALSE)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{size}{the minimum file size in megabytes to track}

\item{new}{logical, if TRUE will only return untracked files, if FALSE will return all files above the size limit}
}
\value{
A vector of relative file paths to the large files
}
\description{
Identify large files that should be tracked with Box LFS
}
\examples{
tmp <- blfs:::create_test_repo(box_lfs=TRUE, examples = FALSE,
  source_dir = system.file("extdata", package = "blfs"))

#testing files are quite small and don't show up
check_files_blfs(tmp)

#but they do if we change the size
check_files_blfs(tmp, size=0.0002)

#they're already tracked, so if we set new to TRUE we don't see them
check_files_blfs(tmp, size=0.0002, new=TRUE)

unlink(tmp, recursive = TRUE)
}
