% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone-repo.R
\name{clone_repo_blfs}
\alias{clone_repo_blfs}
\title{Get files after cloning a repo that uses Box LFS}
\usage{
clone_repo_blfs(dir = NULL, download = NULL, boxdrive = TRUE)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{download}{the file path to the download directory (if NULL will default to the Box Drive path)}

\item{boxdrive}{logical, should Box drive be used?}
}
\value{
Prompts user to download the files from Box and then places them in the correct location in the \code{dir} folder
}
\description{
If you want to clone a Github repository that uses Box LFS, you'll need to manually download the tracked files from Box so
you have them in your repository. This code will prompt you to download the Box files, then place them in the correct files
in your repository using the .boxtracker files.
}
\examples{
 #create temp dir to modify files cleanly
  tmp <- blfs:::create_test_repo(box_lfs=TRUE, examples = FALSE,
  source_dir = system.file("extdata", package = "blfs"))

  box_tmp <- blfs:::create_test_boxdrive(source_dir = system.file("extdata", package = "blfs"))

  #put in new file path to "box"
   blfs:::add_box_loc(box_tmp, dir=tmp, type="path")

  clone_repo_blfs(dir=tmp, download=box_tmp)

  #remove temp dirs
  unlink(box_tmp, recursive = TRUE)
  unlink(tmp, recursive = TRUE)
}
