% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{create_download_file}
\alias{create_download_file}
\title{Fake a file that needs to be downloaded}
\usage{
create_download_file(name, dir)
}
\arguments{
\item{name}{the relative file path to the file to fake download}

\item{dir}{the fake project repository}
}
\value{
saves the updated box tracker file in the fake project
}
\description{
Fake a file that needs to be downloaded
}
