% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{create_updated_file}
\alias{create_updated_file}
\title{Fake a file that needs to be updated}
\usage{
create_updated_file(name, dir)
}
\arguments{
\item{name}{the relative file path to the file to fake update}

\item{dir}{the fake project repository}
}
\value{
saves the updated box tracker file in the fake project
}
\description{
Fake a file that needs to be updated
}
