% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtracker.R
\name{get.boxtracker}
\alias{get.boxtracker}
\title{Get and format info for boxtracker file}
\usage{
get.boxtracker(file, dir = NULL)
}
\arguments{
\item{file}{the relative path to the file to track}

\item{dir}{the file path to the file directory}
}
\value{
a data.frame with 1 row and 5 columns:
\itemize{
\item file_path: the relative path to the file being tracked
\item box_link: the web link to the folder containing the file
\item box_path: the box drive path to the folder containing the file
\item size_MB: the size of the file in megabytes
\item last_modified: the date and time the file was last modified
\item last_changed: the date and time the file was last changed
}
}
\description{
Pulls and formats the data from a file for the boxtracker file, but does not write to the file (see \link[blfs]{write.boxtracker})
}
\examples{
get.boxtracker("example-files/large-file1.txt", fs::path_package("extdata", package = "blfs"))
}
