% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_name}
\alias{get_file_name}
\title{Get the file path associated with a boxtracker}
\usage{
get_file_name(dir = NULL, tracker)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{tracker}{the name of the file with the .boxtracker extension (should be a hash)}
}
\value{
the file path associated with a hash based .boxtracker file.
}
\description{
Get the file path associated with a boxtracker
}
\examples{
get_file_name("1678f723cb201eb3f9996c01a481dd0e.boxtracker",
dir=fs::path_package("extdata", package = "blfs"))
}
