% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tracker_name}
\alias{get_tracker_name}
\title{Creates a unique file name for each boxtracker}
\usage{
get_tracker_name(file)
}
\arguments{
\item{file}{the file to be tracked}
}
\value{
the tracker name as a unique hash based on the file name with the extension ".boxtracker"
}
\description{
Uses the \link[digest]{digest} function to generate a unique serialized hash code for each file being tracked.
}
\examples{
get_tracker_name("test-file1.txt")
get_tracker_name("another-folder/another_file.txt")
}
