% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutifile.R
\name{is.multifile}
\alias{is.multifile}
\title{Determines if a file is a multifile}
\usage{
is.multifile(file, names = FALSE)
}
\arguments{
\item{file}{the path to the multifile}

\item{names}{if TRUE, provides names and ext of the files within multifile}
}
\value{
TRUE if multifile, otherwise FALSE
}
\description{
Determines if a file is a multifile
}
\examples{
file <- "inst/extdata/example-files/example-shp.shp"
blfs:::is.multifile(file)

file <- "inst/extdata/example-files/large-file1.txt"
blfs:::is.multifile(file)

}
