% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain-blfs.R
\name{move_file_blfs}
\alias{move_file_blfs}
\title{Moved downloaded files to the correct file path in the project}
\usage{
move_file_blfs(hash_file, dir = NULL, download = NULL)
}
\arguments{
\item{hash_file}{the file to track, should match the same hash name as the tracker associated with the file}

\item{dir}{the file path to the file directory}

\item{download}{the file path to the download directory}
}
\value{
Copies files from the download folder to the project directory in the correct subfolder location based on the .boxtracker file.
}
\description{
Moved downloaded files to the correct file path in the project
}
\note{
If download is not supplied function assumes it is \code{file.path(fs::path_home(), "Downloads")}
}
\examples{
#create temp dir to modify files cleanly
  tmp <- blfs:::create_test_repo(box_lfs=TRUE, examples = FALSE,
  source_dir = system.file("extdata", package = "blfs"))
  withr::defer(unlink(tmp, recursive = TRUE), envir = parent.frame())

  download <- blfs:::create_test_boxdrive(source_dir = system.file("extdata", package = "blfs"))
  withr::defer(unlink(download, recursive = TRUE), envir = parent.frame())

#move file from download
  move_file_blfs("1678f723cb201eb3f9996c01a481dd0e.txt",
  dir=tmp, download=download)

  move_file_blfs("3f80f3c380f48192c6fcd63a08813c49.zip",
  dir=tmp, download=download)

  unlink(download, recursive = TRUE)
  unlink(tmp, recursive = TRUE)
}
