% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prj_name}
\alias{prj_name}
\title{Get project name from a github repository}
\usage{
prj_name(dir)
}
\arguments{
\item{dir}{the file path to the file directory}
}
\value{
a character giving the project name based on a GitHub repository name
}
\description{
Removes the WWS-Node#- from the repo name as it's not needed on Box
}
\examples{
prj_name("~/Documents/WWS-Node1-TEST-example-repo")
}
