% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-repo.R
\name{pull_repo_blfs}
\alias{pull_repo_blfs}
\title{Check for updated Box LFS files when pulling a GitHub repository}
\usage{
pull_repo_blfs(dir = NULL, download = NULL, boxdrive = TRUE)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{download}{the file path to the download directory (if NULL will default to the Box Drive path)}

\item{boxdrive}{logical, should Box drive be used?}
}
\value{
Checks for:
\itemize{
\item tracked files in the local directory that are newer than the tracker and prompts upload
\item tracked files that are new or newer on Box than the local directory and prompts download
}
}
\description{
Run AFTER pulling files from GitHub.
}
\details{
When you pull the updates from a GitHub repository that has large files tracked with Box LFS you also
may need to update those files. This function will check the status of the tracked files in your local
repository and determine if there are any files that uploaded or downloaded to maintain the current version.
}
\examples{
 #create temp dir to modify files cleanly
  tmp <- blfs:::create_test_repo(box_lfs=TRUE, examples = FALSE,
  source_dir = system.file("extdata", package = "blfs"))

  box_tmp <- blfs:::create_test_boxdrive(source_dir = system.file("extdata", package = "blfs"))

  #put in new file path to "box"
   blfs:::add_box_loc(box_tmp, dir=tmp, type="path")

pull_repo_blfs(dir=tmp, download=box_tmp)

 unlink(box_tmp, recursive = TRUE)
 unlink(tmp, recursive = TRUE)
}
