% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtracker.R
\name{read.boxtracker}
\alias{read.boxtracker}
\title{Get file location or other data from boxtracker}
\usage{
read.boxtracker(tracker, dir = NULL, return = "all")
}
\arguments{
\item{tracker}{the name of the file with the .boxtracker extension}

\item{dir}{the file path to the file directory}

\item{return}{the column to return. options are: file_path, box_link, size_MB, last_modified, last_changed}
}
\value{
\itemize{
\item if \code{return} is "all" will return a data.frame
\item otherwise will return a vector of length one with the column value
}
}
\description{
The .boxtracker files stores the relative location of the file, the box link, and file info (size, dates last modified and changed).
This function is used to easily read and extract data from those files.
}
\examples{
read.boxtracker("large-file1", fs::path_package("extdata", package = "blfs"))
read.boxtracker("large-file1", fs::path_package("extdata", package = "blfs"), return = "size_MB")

}
