% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readme_msg}
\alias{readme_msg}
\title{Helper to write message about Box LFS to README.md}
\usage{
readme_msg(readme)
}
\value{
Lines to add to the readme file informing user that Box LFS is being used
}
\description{
Helper to write message about Box LFS to README.md
}
\examples{
readme_msg()
}
