% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-blfs.R
\name{track_blfs}
\alias{track_blfs}
\title{Start tracking a file with Box LFS}
\usage{
track_blfs(file, dir = NULL)
}
\arguments{
\item{file}{the relative path to the file to track}

\item{dir}{the file path to the file directory}
}
\value{
the name of the file being tracked to use in a warning message
}
\description{
Creates a .boxtracker file with the file information, adds the file to .gitignore, and copy to upload folder so it's easy to but on Box.
If it's a multifile, will add all files to .gitignore and copy a .zip of all files to the upload folder.
}
\examples{
track_blfs("example-files/large-file1.txt", fs::path_package("extdata", package = "blfs"))
track_blfs("example-files/example-shp", fs::path_package("extdata", package = "blfs"))
}
