% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtracker.R
\name{write.boxtracker}
\alias{write.boxtracker}
\title{Write boxtracker file}
\usage{
write.boxtracker(file, dir = NULL)
}
\arguments{
\item{file}{the relative path to the file to track}

\item{dir}{the file path to the file directory}
}
\value{
Saves a boxtracker file to dir/box-lfs with the form file.boxtracker. See \link[blfs]{get.boxtracker} for details on the file structure.
}
\description{
Creates the boxtracker file with a .boxtracker extension to track a file with blfs. If the file is being created for the first time,
it will create a new line in the path-hash.csv file which links the hash names to the file paths.
}
\examples{
write.boxtracker("example-files/large-file1.txt", fs::path_package("extdata", package = "blfs"))
write.boxtracker("example-files/example-shp.shp", fs::path_package("extdata", package = "blfs"))
}
