% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blfs-package.R
\docType{package}
\name{blfs-package}
\alias{blfs}
\alias{blfs-package}
\title{blfs: Track Large Project Files Stored on Box}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Katie Wampler \email{katie.wampler@oregonstate.edu} (\href{https://orcid.org/0000-0002-5584-7554}{ORCID})

}
\keyword{internal}
