% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_blfs}
\alias{check_blfs}
\title{Check if Box LFS is being used on the project directory}
\usage{
check_blfs(dir = NULL)
}
\arguments{
\item{dir}{the file path to the file directory}
}
\value{
A value of TRUE if Box LFS is being used or FALSE if it is not
}
\description{
Check if Box LFS is being used on the project directory
}
\examples{
check_blfs(fs::path_package("extdata", package = "blfs"))
}
