% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone-repo.R
\name{clone_repo_blfs}
\alias{clone_repo_blfs}
\title{Get files after cloning a repo that uses box-lfs}
\usage{
clone_repo_blfs(dir = NULL, download = NULL)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{download}{the file path to the download directory}
}
\value{
Prompts user to download the files from Box and then places them in the correct location in the \code{dir} folder
}
\description{
If you want to clone a Github repository that uses box-lfs, you'll need to manually download the tracked files from Box so
you have them in your repository. This code will prompt you to download the Box files, then place them in the correct files
in your repository using the .boxtracker files.
}
\examples{
tmp <- withr::local_tempdir()

#move just tracker files in, similar to cloning a repo from github
file.copy(fs::path_package("extdata/box-lfs", package = "blfs"), tmp, recursive=TRUE)
download <- fs::path_package("extdata", package = "blfs") #example zip is here

clone_repo_blfs(dir=tmp, download=download)
}
