% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-blfs.R
\name{init_blfs}
\alias{init_blfs}
\title{Set up the structure for box-lfs}
\usage{
init_blfs(dir = NULL)
}
\arguments{
\item{dir}{the file path to the file directory}
}
\value{
Creates a box-lfs in the directory, with a upload folder nested inside.
Also creates a .gitignore file it it doesn't already exists and adds the upload folder to it.
}
\description{
Creates the file structure for tracking large files with box. Adds the upload folder to .gitignore.
If used on a folder that already has the file stucture set up it will run without errors.
}
\examples{
init_blfs(fs::path_package("extdata", package = "blfs"))
}
