% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain-blfs.R
\name{move_file_blfs}
\alias{move_file_blfs}
\title{Moved downloaded files to the correct file path in the project}
\usage{
move_file_blfs(file, dir = NULL, download = NULL)
}
\arguments{
\item{file}{the file to track, does not need to have relative path}

\item{dir}{the file path to the file directory}

\item{download}{the file path to the download directory}
}
\value{
Copies files from the download folder to the project directory in the correct subfolder location based on the .boxtracker file.
}
\description{
Moved downloaded files to the correct file path in the project
}
\note{
If download is not supplied function assumes it is \code{file.path(fs::path_home(), "Downloads")}
}
\examples{
#returns false because file doesn't exist in downloads folder
move_file_blfs("large-file1.txt", dir="ext-data")
}
