% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_repo_blfs.R
\name{push_repo_blfs}
\alias{push_repo_blfs}
\title{Check for large files before pushing a Git repository}
\usage{
push_repo_blfs(dir = NULL, size = 10)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{size}{the minimum file size in megabytes to track}
}
\value{
Identifies any new or modified files, updates the .boxtracker file and prompts user to upload to Box
}
\description{
Run BEFORE pushing to GitHub.
}
\details{
If you try to push files larger than 100 MB to GitHub you will receive an error. This function will identify files
above the size threshold. If they are already tracked, they will be checked for updates. If there are new files they
will be tracked and the user prompted to upload to Box.
}
\examples{
push_repo_blfs(fs::path_package("extdata", package = "blfs"))
}
