% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain-blfs.R
\name{update_blfs}
\alias{update_blfs}
\title{Update files that are tracked by Box LFS}
\usage{
update_blfs(file, dir = NULL)
}
\arguments{
\item{file}{the relative path to the file to track}

\item{dir}{the file path to the file directory}
}
\value{
\itemize{
\item If the boxtracker shows a newer file, it returns "download"
\item If the file shows a newer file, it returns "upload"
\item If the boxtracker and file are up to date with each other it will return nothing
}
}
\description{
Checks for differences in the time modified between the boxtracker and the file.
}
\examples{
update_blfs("example-files/large-file1.txt", fs::path_package("extdata", package = "blfs"))
}
